Minor fixes
It is possible, under rare (ready broken) circumstances that the deathLocations will be null and not an empty hashmap.
Also, fixed issues where the force player respawn command wasn't correctl updating that players death data. Added a respawningPlayer method to death data to handle any specific requirements internally (for now, it' sjust resetting the respawn time to the current time)